Folder Authentication, refactored AuthValidator.cs
