+++ SmartSaver/Authentication/AuthValidator.cs
@@ -3,8 +3,9 @@ using System.Text.RegularExpressions;
 using System.Globalization;
 using System.Linq;
 using ePiggy.forms;
+using ePiggy.utilities;
 
-namespace DataManager
+namespace ePiggy.Authentication
 {
     public static class AuthValidator
     {
@@ -59,17 +60,19 @@ namespace DataManager
                     var idn = new IdnMapping();
 
                     // Pull out and process domain name (throws ArgumentException on invalid)
-                    string domainName = idn.GetAscii(match.Groups[2].Value);
+                    var domainName = idn.GetAscii(match.Groups[2].Value);
 
                     return match.Groups[1].Value + domainName;
                 }
             }
             catch (RegexMatchTimeoutException e)
             {
+                ExceptionHandler.Log(e.ToString());
                 return false;
             }
             catch (ArgumentException e)
             {
+                ExceptionHandler.Log(e.ToString());
                 return false;
             }
