+++ SmartSaver/Authentication/HashingProcessor.cs
