+++ SmartSaver/DataManager/GraphDrawer.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Windows.Forms;
 using ImageChartsLib;
 
-namespace ePiggy.backend
+namespace ePiggy.DataManager
 {
     class GraphDrawer
     {
