+++ SmartSaver/DataManager/OfferData.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DataManager
+namespace ePiggy.DataManager
 {
     public class OfferData
     {
